add index js re-exports
