--- controllers/auth.js
@@ -1,84 +1,0 @@
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
-
-const { User } = require('../models/user');
-const { controllerWrapper, HttpError } = require('../helpers');
-
-const { SECRET_KEY } = process.env;
-
-const register = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (user) {
-    throw HttpError(409, 'Email in use');
-  }
-  const createHashPassword = await bcrypt.hash(password, 10);
-
-  const newUser = await User.create({
-    ...req.body,
-    password: createHashPassword,
-  });
-  res.status(201).json({
-    user: {
-      email: newUser.email,
-      subscription: 'starter',
-    },
-  });
-};
-
-const login = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (!user) {
-    throw HttpError(401, 'Email or password is wrong');
-  }
-
-  const comparePassword = await bcrypt.compare(password, user.password);
-  if (!comparePassword) {
-    throw HttpError(401, 'Email or password is wrong');
-  }
-
-  const payload = {
-    id: user._id,
-  };
-
-  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: '23h' });
-  await User.findByIdAndUpdate(user._id, { token });
-
-
-  res.json({
-    token,
-    user: {
-      email: user.email,
-      subscription: 'starter',
-    },
-  });
-};
-
-const getCurrent = async (req, res) => {
-  const { email, subscription } = req.user;
-  res.json({ email, subscription });
-}
-
-const logout = async (req, res) => {
-  const { _id } = req.user;
-  await User.findByIdAndUpdate(_id, { token: '' });
-  res.status(204);
-};
-
-const updateSubscription = async (req, res) => {
-  const { _id } = req.user;
-  const result = await User.findByIdAndUpdate(_id, req.body, { new: true });
-  if (!result) {
-    throw HttpError(404, 'Not found');
-  }
-  res.json(result);
-}
-
-module.exports = {
-  register: controllerWrapper(register),
-  login: controllerWrapper(login),
-  getCurrent: controllerWrapper(getCurrent),
-  logout: controllerWrapper(logout),
-  updateSubscription: controllerWrapper(updateSubscription),
-};
