+++ controllers/ctrlAuth/getCurrentUser.js
@@ -1,0 +1,6 @@
+const getCurrent = async (req, res) => {
+  const { email, subscription } = req.user;
+  res.json({ email, subscription });
+};
+
+module.exports = getCurrent;
