+++ controllers/ctrlAuth/register.js
@@ -1,0 +1,25 @@
+const bcrypt = require('bcrypt');
+const { User } = require('../../models/user');
+const { HttpError } = require('../../helpers');
+
+const register = async (req, res) => {
+  const { email, password } = req.body;
+  const user = await User.findOne({ email });
+  if (user) {
+    throw HttpError(409, 'Email in use');
+  }
+  const createHashPassword = await bcrypt.hash(password, 10);
+
+  const newUser = await User.create({
+    ...req.body,
+    password: createHashPassword,
+  });
+  res.status(201).json({
+    user: {
+      email: newUser.email,
+      subscription: 'starter',
+    },
+  });
+};
+
+module.exports = register;
