+++ controllers/ctrlContacts/onAddNewContact.js
@@ -1,0 +1,9 @@
+const { Contact } = require('../../models/contact');
+
+const onAddNewContact = async (req, res) => {
+  const { _id: owner } = req.user;
+  const result = await Contact.create({ ...req.body, owner });
+  res.status(201).json(result);
+};
+
+module.exports = onAddNewContact;
