--- routes/api/index.js
@@ -1,5 +1,0 @@
-const router = require("./auth");
-
-module.exports = {
-  router,
-}
